中文版注释

+++ ch06/hyperparameter_optimization.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 import sys, os
-sys.path.append(os.pardir)  # 親ディレクトリのファイルをインポートするための設定
+sys.path.append(os.pardir)  # 为了导入父目录的文件而进行的设定
 import numpy as np
 import matplotlib.pyplot as plt
 from dataset.mnist import load_mnist
@@ -10,13 +10,13 @@ from common.trainer import Trainer
 
 (x_train, t_train), (x_test, t_test) = load_mnist(normalize=True)
 
-# 高速化のため訓練データの削減
+# 为了实现高速化，减少训练数据
 x_train = x_train[:500]
 t_train = t_train[:500]
 
-# 検証データの分離
+# 分割验证数据
 validation_rate = 0.20
-validation_num = int(x_train.shape[0] * validation_rate)
+validation_num = x_train.shape[0] * validation_rate
 x_train, t_train = shuffle_dataset(x_train, t_train)
 x_val = x_train[:validation_num]
 t_val = t_train[:validation_num]
@@ -35,12 +35,12 @@ def __train(lr, weight_decay, epocs=50):
     return trainer.test_acc_list, trainer.train_acc_list
 
 
-# ハイパーパラメータのランダム探索======================================
+# 超参数的随机搜索======================================
 optimization_trial = 100
 results_val = {}
 results_train = {}
 for _ in range(optimization_trial):
-    # 探索したハイパーパラメータの範囲を指定===============
+    # 指定搜索的超参数的范围===============
     weight_decay = 10 ** np.random.uniform(-8, -4)
     lr = 10 ** np.random.uniform(-6, -2)
     # ================================================
@@ -51,7 +51,7 @@ for _ in range(optimization_trial):
     results_val[key] = val_acc_list
     results_train[key] = train_acc_list
 
-# グラフの描画========================================================
+# 绘制图形========================================================
 print("=========== Hyper-Parameter Optimization Result ===========")
 graph_draw_num = 20
 col_num = 5
